Finished implementation on search bar. Will search and index through one or many input fields.
