--- src/main/webapp/WEB-INF/ads/update.jsp
@@ -1,9 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>Title</title>
-</head>
-<body>
-
-</body>
-</html>
